requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow>=11.3.0
pillow-heif>=1.1.0
google-cloud-vision==3.4.4
openai==1.3.3